Arrumando conflitos e finalizando a issue de balance sheets por company id

+++ src/main/java/com/stocksapi/model/BalanceSheet.java
@@ -10,8 +10,8 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-import java.math.BigInteger;
-import java.math.BigInteger;
+import java.math.BigDecimal;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "balance_sheets", schema = "public")
@@ -24,48 +24,48 @@ public class BalanceSheet {
     private Companies companies;
 
     @Column(name = "net_revenue")
-    private BigInteger netRevenue;
+    private BigDecimal netRevenue;
 
     @Column(name = "costs")
-    private BigInteger costs;
+    private BigDecimal costs;
 
     @Column(name = "gross_profit")
-    private BigInteger grossProfit;
+    private BigDecimal grossProfit;
 
     @Column(name = "net_profit")
-    private BigInteger netProfit;
+    private BigDecimal netProfit;
 
     @Column(name = "ebitda")
-    private BigInteger ebitda;
+    private BigDecimal ebitda;
 
     @Column(name = "ebit")
-    private BigInteger ebit;
+    private BigDecimal ebit;
 
     @Column(name = "taxes")
-    private BigInteger taxes;
+    private BigDecimal taxes;
 
     @Column(name = "gross_debt")
-    private BigInteger grossDebt;
+    private BigDecimal grossDebt;
 
     @Column(name = "net_debt")
-    private BigInteger netDebt;
+    private BigDecimal netDebt;
 
     @Column(name = "equity")
-    private BigInteger equity;
+    private BigDecimal equity;
 
     @Column(name = "cash")
-    private BigInteger cash;
+    private BigDecimal cash;
 
     @Column(name = "assets")
-    private BigInteger assets;
+    private BigDecimal assets;
 
     @Column(name = "liabilities")
-    private BigInteger liabilities;
+    private BigDecimal liabilities;
 
     public BalanceSheet() {
     }
 
-    public BalanceSheet(BigInteger assets, BigInteger cash, Companies companies, BigInteger costs, BigInteger ebit, BigInteger ebitda, BigInteger equity, BigInteger grossDebt, BigInteger grossProfit, BalanceSheetId id, BigInteger liabilities, BigInteger netDebt, BigInteger netProfit, BigInteger netRevenue, BigInteger taxes) {
+    public BalanceSheet(BigDecimal assets, BigDecimal cash, Companies companies, BigDecimal costs, BigDecimal ebit, BigDecimal ebitda, BigDecimal equity, BigDecimal grossDebt, BigDecimal grossProfit, BalanceSheetId id, BigDecimal liabilities, BigDecimal netDebt, BigDecimal netProfit, BigDecimal netRevenue, BigDecimal taxes) {
         this.assets = assets;
         this.cash = cash;
         this.companies = companies;
@@ -83,11 +83,11 @@ public class BalanceSheet {
         this.taxes = taxes;
     }
 
-    public BigInteger getAssets() {
+    public BigDecimal getAssets() {
         return assets;
     }
 
-    public BigInteger getCash() {
+    public BigDecimal getCash() {
         return cash;
     }
 
@@ -95,27 +95,27 @@ public class BalanceSheet {
         return companies;
     }
 
-    public BigInteger getCosts() {
+    public BigDecimal getCosts() {
         return costs;
     }
 
-    public BigInteger getEbit() {
+    public BigDecimal getEbit() {
         return ebit;
     }
 
-    public BigInteger getEbitda() {
+    public BigDecimal getEbitda() {
         return ebitda;
     }
 
-    public BigInteger getEquity() {
+    public BigDecimal getEquity() {
         return equity;
     }
 
-    public BigInteger getGrossDebt() {
+    public BigDecimal getGrossDebt() {
         return grossDebt;
     }
 
-    public BigInteger getGrossProfit() {
+    public BigDecimal getGrossProfit() {
         return grossProfit;
     }
 
@@ -123,23 +123,23 @@ public class BalanceSheet {
         return id;
     }
 
-    public BigInteger getLiabilities() {
+    public BigDecimal getLiabilities() {
         return liabilities;
     }
 
-    public BigInteger getNetDebt() {
+    public BigDecimal getNetDebt() {
         return netDebt;
     }
 
-    public BigInteger getNetProfit() {
+    public BigDecimal getNetProfit() {
         return netProfit;
     }
 
-    public BigInteger getNetRevenue() {
+    public BigDecimal getNetRevenue() {
         return netRevenue;
     }
 
-    public BigInteger getTaxes() {
+    public BigDecimal getTaxes() {
         return taxes;
     }
 }
